amended and restated articles of association
Previously, International Protectors and Advisors Ltd. (a wholly-owned subsidiary of International Corporation Services Ltd., itself an affiliate of Travers Thorp Alberga) exercised the rights of both the Founder and the Supervisor under the constitutional documents of the Foundation. For internal administrative reasons and to provide enhanced governance, TTA Founders Limited now exercises the rights of the Founder and TTA Supervisors Limited now exercises the rights of the Supervisor, with respect to the Foundation. TTA Founders Limited and TTA Supervisors Limited are both wholly-owned subsidiaries of International Corporation Services Ltd., which is licensed and regulated by the Cayman Islands Monetary Authority.